Solve '07-flex-layout-2,' finish Flex challenges
This one had its challenges, but mostly with centering the header,
as it was simply text inside of a div and it was harder to target that
text without it being an element. Their solution doesn't seem to have
the problems I did, but making them paragraphs worked just fine.
Everything else came without major challenges.